+++ src/main/java/net/codejava/repository/SymptomRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import net.codejava.entity.Symptom;
 
-public interface SymptomRepository extends JpaRepository<Symptom, Long>{
+public interface SymptomRepository extends JpaRepository<Symptom, Integer>{
 
 }
