+++ src/main/java/net/codejava/repository/WeightRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import net.codejava.entity.Weight;
 
-public interface WeightRepository extends JpaRepository<Weight, Long>{
+public interface WeightRepository extends JpaRepository<Weight, Integer>{
 
 }
